requirements.txt
pillow
aiogram